move storage interface to storage folder.
